pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bess-model"
version = "0.1.0"
description = "Scalable BESS sizing and simulation framework for minute-level renewable data."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "Flask>=3.1.0",
  "numba>=0.60.0",
  "numpy>=2.0.0",
  "polars>=1.22.0",
  "PyYAML>=6.0.2",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
]

[project.scripts]
bess-model = "bess_model.main:main"
bess-model-web = "bess_model.web.app:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["bess_model*"]

[tool.setuptools.package-data]
"bess_model.web" = ["templates/*.html", "static/*.css", "static/*.js"]

[tool.pytest.ini_options]
testpaths = ["tests"]
